requirements.txt
aiohttp==3.5.4
alembic==1.0.8
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cryptography==2.6.1
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
future==0.17.1
idna==2.8
idna-ssl==1.1.0
isort==4.3.15
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.8
MarkupSafe==1.1.1
mccabe==0.6.1
multidict==4.5.2
pkg-resources==0.0.0
pprint==0.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
py-postgresql==1.2.1
pycparser==2.19
pylint==2.3.1
pylint-flask==0.6
pylint-plugin-utils==0.5
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
python-telegram-bot==11.1.0
requests==2.21.0
six==1.12.0
SQLAlchemy==1.3.1
telegram==0.0.1
telepot==12.7
timer3==0.1.2
typed-ast==1.3.1
typing-extensions==3.7.2
urllib3==1.24.1
Werkzeug==0.15.0
wrapt==1.11.1
yarl==1.3.0
